Run series of tests and write 1 data row for each to csv
Our end goal is to measure how PTracker scales with increasing load, so we don't need to plot the totals metrics except at the end of the test. This also requires bounding the lifetime of the worker threads, but it allows us to remove the printer thread concept and instead write data synchronously after each test.